ready front v1.0

+++ frontend/src/components/cards/BankCard.tsx
@@ -13,24 +13,28 @@ interface BankCardProps {
 }
 
 function BankCard({ id, bankName, balance, nextBillingDate, billingBalance, onDelete, onOpenModal }: BankCardProps) {
-
+    
     const handleDelete = async () => {
-        try {
-            await BankAccountsService.deleteBankAccounts(id);
-            console.log('Conta bancária deletada com sucesso');
-            // Se onDelete for definido, chame-o para executar qualquer ação adicional necessária após a exclusão
-            if (onDelete) {
-                onDelete();
+        const confirmed = window.confirm('Tem certeza de que deseja excluir esta conta bancária?');
+        if (confirmed) {
+            try {
+                await BankAccountsService.deleteBankAccounts(id);
+                console.log('Conta bancária deletada com sucesso');
+                // Se onDelete for definido, chame-o para executar qualquer ação adicional necessária após a exclusão
+                if (onDelete) {
+                    onDelete();
+                }
+            } catch (error) {
+                console.error('Erro ao deletar conta bancária:', error);
             }
-        } catch (error) {
-            console.error('Erro ao deletar conta bancária:', error);
         }
     };
+
 
     return (
         <div className="bg-gray-50 shadow-lg rounded-lg overflow-hidden max-w-80 w-72 m-4">
             <div className="bg-green-500 text-white px-4 py-2 p-8">
-                <h2 className="text-xl font-semibold">Bank {bankName}</h2>
+                <h2 className="text-xl font-semibold">{bankName}</h2>
             </div>
             <div className="p-4">
                 <div className="mb-2">
